# cmd/apiserver/docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  model.GitFileAppendManualReq:
    properties:
      gitLink:
        type: string
    type: object
  model.GitFileDTO:
    properties:
      failedTimes:
        type: integer
      filePath:
        type: string
      gitLink:
        type: string
      lastSuccess:
        type: string
      message:
        type: string
      success:
        type: boolean
      takeStorage:
        type: integer
      takeTimeMs:
        type: integer
      updateTime:
        type: string
    type: object
  model.GitFileStatisticsResultDTO:
    properties:
      fail:
        type: integer
      neverSuccess:
        type: integer
      success:
        type: integer
      total:
        type: integer
    type: object
  model.GitFileStatusResp:
    properties:
      collector:
        $ref: '#/definitions/rpc.StatusResp'
      gitFile:
        $ref: '#/definitions/model.GitFileStatisticsResultDTO'
    type: object
  model.GitHubCallbackResp:
    properties:
      token:
        type: string
    type: object
  model.GitHubClientIDResp:
    properties:
      clientId:
        type: string
    type: object
  model.PageDTO-model_GitFileDTO:
    properties:
      count:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.GitFileDTO'
        type: array
      start:
        type: integer
      total:
        type: integer
    type: object
  model.PageDTO-model_RankingResultDTO:
    properties:
      count:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.RankingResultDTO'
        type: array
      start:
        type: integer
      total:
        type: integer
    type: object
  model.PageDTO-model_ResultDTO:
    properties:
      count:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.ResultDTO'
        type: array
      start:
        type: integer
      total:
        type: integer
    type: object
  model.RankingResultDTO:
    properties:
      distDetail:
        items:
          $ref: '#/definitions/model.ResultDistDetailDTO'
        type: array
      distroScore:
        type: number
      gitDetail:
        items:
          $ref: '#/definitions/model.ResultGitMetadataDTO'
        type: array
      gitScore:
        type: number
      langDetail:
        items:
          $ref: '#/definitions/model.ResultLangDetailDTO'
        type: array
      langScore:
        type: number
      link:
        type: string
      ranking:
        type: integer
      score:
        type: number
      scoreID:
        type: integer
      updateTime:
        type: string
    type: object
  model.ResultDTO:
    properties:
      distDetail:
        items:
          $ref: '#/definitions/model.ResultDistDetailDTO'
        type: array
      distroScore:
        type: number
      gitDetail:
        items:
          $ref: '#/definitions/model.ResultGitMetadataDTO'
        type: array
      gitScore:
        type: number
      langDetail:
        items:
          $ref: '#/definitions/model.ResultLangDetailDTO'
        type: array
      langScore:
        type: number
      link:
        type: string
      score:
        type: number
      scoreID:
        type: integer
      updateTime:
        type: string
    type: object
  model.ResultDistDetailDTO:
    properties:
      count:
        type: integer
      impact:
        type: number
      pageRank:
        type: number
      type:
        type: integer
      updateTime:
        type: string
    type: object
  model.ResultGitMetadataDTO:
    properties:
      commitFrequency:
        type: number
      contributorCount:
        type: integer
      createdSince:
        type: string
      language:
        items:
          type: string
        type: array
      license:
        items:
          type: string
        type: array
      orgCount:
        type: integer
      updateTime:
        type: string
      updatedSince:
        type: string
    type: object
  model.ResultLangDetailDTO:
    properties:
      depCount:
        type: integer
      langEcoImpact:
        type: number
      type:
        type: integer
      updateTime:
        type: string
    type: object
  model.ToolArgDTO:
    properties:
      default: {}
      description:
        type: string
      name:
        description: Name is the name of the argument.
        type: string
      type:
        type: string
    type: object
  model.ToolCreateInstanceReq:
    properties:
      args:
        additionalProperties: {}
        type: object
      toolId:
        type: string
    type: object
  model.ToolDTO:
    properties:
      args:
        items:
          $ref: '#/definitions/model.ToolArgDTO'
        type: array
      description:
        type: string
      id:
        description: ID is the unique identifier for the toolset.
        type: string
      name:
        type: string
    type: object
  model.ToolInstanceDTO:
    properties:
      id:
        type: string
      startTime:
        type: string
      tool:
        $ref: '#/definitions/model.ToolDTO'
    type: object
  model.UserInfoResp:
    properties:
      policy:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  rpc.RunningTaskDTO:
    properties:
      link:
        type: string
      progress:
        type: string
      start:
        type: string
    type: object
  rpc.StatusResp:
    properties:
      currentTasks:
        items:
          $ref: '#/definitions/rpc.RunningTaskDTO'
        type: array
      isRunning:
        type: boolean
      pendingTasks:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
  /admin/gitfiles:
    get:
      parameters:
      - description: Git link
        in: query
        name: link
        type: string
      - description: 'Filter, 0: no filter, 1: success, 2: fail, 3: never success'
        in: query
        name: filter
        type: integer
      - description: Skip count
        in: query
        name: skip
        type: integer
      - description: Take count
        in: query
        name: take
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PageDTO-model_GitFileDTO'
      summary: Get Git File List
  /admin/gitfiles/manual:
    post:
      consumes:
      - application/json
      parameters:
      - description: Append manual request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.GitFileAppendManualReq'
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Append to collector manual list
  /admin/gitfiles/start:
    post:
      description: Starts the Git File collection process
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Start Git File Collector
  /admin/gitfiles/status:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GitFileStatusResp'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get git files statistics and collector status
  /admin/gitfiles/stop:
    post:
      description: Stops the Git File collection process
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Stop Git File Collector
  /admin/session/github/callback:
    get:
      description: Handles the GitHub OAuth callback and returns JWT token if user
        is authorized
      parameters:
      - description: GitHub OAuth Code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GitHubCallbackResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: GitHub OAuth callback
      tags:
      - admin
  /admin/session/github/clientid:
    get:
      description: Get github client id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GitHubClientIDResp'
      summary: Get github client id
      tags:
      - admin
  /admin/session/userinfo:
    get:
      description: Returns the authenticated user's username and policy
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserInfoResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user information
      tags:
      - admin
  /admin/toolset/instances:
    get:
      description: 获取所有运行中的工具实例的信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ToolInstanceDTO'
            type: array
      summary: 获取运行中的工具实例列表
      tags:
      - toolset
    post:
      consumes:
      - application/json
      description: 根据工具ID和参数创建并运行工具实例
      parameters:
      - description: 工具实例创建参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ToolCreateInstanceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ToolInstanceDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: 创建工具实例
      tags:
      - toolset
  /admin/toolset/instances/{id}/attach:
    get:
      description: 通过 WebSocket 方式 attach 到指定工具实例
      parameters:
      - description: 实例ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: WebSocket 连接已建立
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: WebSocket 连接工具实例
      tags:
      - toolset
  /admin/toolset/instances/{id}/log:
    get:
      description: 获取指定工具实例的日志
      parameters:
      - description: 实例ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: 获取工具实例日志
      tags:
      - toolset
  /admin/toolset/list:
    get:
      description: 获取所有可用工具的信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ToolDTO'
            type: array
      summary: 获取工具列表
      tags:
      - toolset
  /histories:
    get:
      consumes:
      - application/json
      description: Get score histories by git link
      parameters:
      - description: Git link
        in: query
        name: link
        required: true
        type: string
      - description: Skip count
        in: query
        name: start
        type: integer
      - description: Take count
        in: query
        name: take
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PageDTO-model_ResultDTO'
      summary: Get score histories
  /rankings:
    get:
      consumes:
      - application/json
      description: Get ranking results, optionally including all details
      parameters:
      - description: Skip count
        in: query
        name: start
        type: integer
      - description: Take count
        in: query
        name: take
        type: integer
      - description: Include details
        in: query
        name: detail
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PageDTO-model_RankingResultDTO'
      summary: Get ranking results
  /results:
    get:
      consumes:
      - application/json
      description: |-
        Search score results by git link
        NOTE: All details are ignored, should use /results/:scoreid to get details
        NOTE: Maxium take count is 1000
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Skip count
        in: query
        name: start
        type: integer
      - description: Take count
        in: query
        name: take
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PageDTO-model_ResultDTO'
      summary: Search score results by git link
  /results/{scoreid}:
    get:
      consumes:
      - application/json
      description: Get score results, including all details by scoreid
      parameters:
      - description: Score ID
        in: path
        name: scoreid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResultDTO'
      summary: Get score results
swagger: "2.0"